+++ test/cypress.config.js
@@ -5,7 +5,7 @@ export default defineConfig({
     experimentalStudio: true,
     experimentalRunAllSpecs: true,
     defaultCommandTimeout: 10000, /* is OK for very slow computers */
-    baseUrl: 'http://localhost:5173/random/path',
+    baseUrl: 'http://localhost:5173/random/path/', /* include trailing slash */
     video: false
   },
 });